Initial work on splitting up sprite
